intelli-credit/ai-services/research-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
textblob==0.18.0.post0
python-dateutil==2.8.2
